agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.14",
    "docker>=7.1.0",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "loguru>=0.7.3",
    "python-consul2>=0.1.5",
    "uvicorn[standard]>=0.35.0",
]
